add dataset, supergraph and model setup
