+++ examples/polypharmacy_gripnet/config
@@ -1,0 +1,18 @@
+from yacs.config import CfgNode
+
+# -----------------------------------------------------------------------------
+# Config definition
+# -----------------------------------------------------------------------------
+
+C = CfgNode()
+
+# -----------------------------------------------------------------------------
+# Dataset
+# -----------------------------------------------------------------------------
+C.DATASET = CfgNode()
+C.DATASET.ROOT = "./data"
+C.DATASET.NAME = "PoSE"
+C.DATASET.URL = "https://github.com/pykale/data/raw/main/graphs/pose_pyg_2.pt"
+
+def get_cfg_defaults():
+    return C.clone()
